Fix -Wc++11-narrowing: /components/exo
This CL was uploaded by git cl split.

R=yhanada@chromium.org

Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2974261
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Yuichiro Hanada <yhanada@chromium.org>
Cr-Commit-Position: refs/heads/master@{#894475}

+++ components/exo/wayland/clients/info.cc
@@ -136,7 +136,7 @@ void AuraOutputScale(void* data,
                      uint32_t scale) {
   Info* info = static_cast<Info*>(data);
 
-  info->next_scales.push_back({flags, scale});
+  info->next_scales.push_back({flags, static_cast<int32_t>(scale)});
 }
 
 void AuraOutputConnection(void* data,
